+++ mlgear/__init__.py
@@ -1,5 +1,5 @@
 from mlgear.cv import run_cv_model
-from mlgear.models import runLGB, runMLP, runLR, runRidge
+from mlgear.models import runLGB, runMLP, runLR, runRidge, get_lgb_feature_importance
 from mlgear.encoders import CategoricalDummyEncoder, ValueCountsEncoder, BayesTargetEncoder, Scaler
 from mlgear.aggregators import mean_diff, std_diff, min_diff, max_diff, nth_smallest, second_min
 from mlgear.metrics import rmse, crps_score, crps_score_, crps_lgb, CRPS_BINS
@@ -8,7 +8,7 @@ from mlgear.utils import print_step, show, chunk, min_max
 
 __all__ = [
     'run_cv_model',
-    'runLGB', 'runMLP', 'runLR', 'runRidge',
+    'runLGB', 'runMLP', 'runLR', 'runRidge', 'get_lgb_feature_importance',
     'CategoricalDummyEncoder', 'ValueCountsEncoder', 'BayesTargetEncoder', 'Scaler',
     'mean_diff', 'std_diff', 'min_diff', 'max_diff', 'nth_smallest', 'second_min',
     'rmse', 'crps_score', 'crps_score_', 'crps_lgb', 'CRPS_BINS',